Alaa

+++ 02-MCAL/ADC/ADC_interface.h
@@ -3,7 +3,7 @@
 
 void ADC_init(void);
 
-uint8_t ADC_u8StartConversionSynch(uint8_t ADC_Channel,uint8_t* u8_Reading);
+uint8_t ADC_u8StartConversionSynch(uint8_t ADC_Channel,uint16_t* u8_Reading);
 
 uint8_t ADC_u8StartConversionASynch(uint8_t ADC_Channel,uint8_t* u8_Reading,void(*NotificFunc)(void));
